make OrdinalMapper top-level and make MapMapper public
